Fix jrt_str_hashcode signed-overflow UB that dropped callers' & 0x7fffffff mask
jrt_str_hashcode computed the Java hashCode polynomial `h = 31 * h + c` in int32_t.
Signed integer overflow is undefined behavior in C, so clang -O2 (under LTO, with the
function inlined into callers) assumed the accumulation never overflows — and a sum of
non-negative bytes starting from 0 with "no overflow" is provably >= 0. It then folded
a caller's `hashCode() & 0x7fffffff` into just `hashCode()`, since the mask is
redundant for a non-negative value. At runtime the hash DOES wrap negative (e.g.
"yellow".hashCode() == -734239628), so the dropped mask let the negative value reach
MiniHashMap.slotFor's array index -> ArrayIndexOutOfBoundsException.

The emitted IR still contained `and i32`, and --no-solver didn't change it, because the
fold happened inside clang under LTO (not the Vire solver); it was also inlining-
dependent (different wrong outputs for equivalent programs) — the classic UB signature.

Fix: accumulate in uint32_t (defined wraparound) and cast to int32_t at the end — the
standard way to compute a Java hashCode in C. clang no longer assumes non-negativity,
so the mask survives.

examples/HashMaps.java is restored to the key "yellow" (negative hashCode), which had
been worked around with the positive-hash "gold"; the 65/65 heap-balance oracle now
covers the negative-hashCode-through-map path as a regression test.

Verified: cargo build clean, cargo test -p vire 63/63, Java heap-balance oracle 65/65
(debug fastjavac rebuilt so the fixed runtime is embedded).

+++ crates/driver/src/runtime.c
@@ -506,11 +506,16 @@ int32_t jrt_obj_hashcode(void *o) {
 
 /* String.hashCode (JLS): s[0]*31^(n-1) + … + s[n-1]. */
 int32_t jrt_str_hashcode(const JStr *s) {
-    int32_t h = 0;
+    /* Accumulate in uint32_t: unsigned arithmetic has DEFINED wraparound. Signed
+     * (int32_t) would be undefined on overflow, and clang -O2/LTO then assumes the
+     * result is non-negative (sum of non-negative terms from 0, "no overflow") and
+     * folds a caller's `hashCode() & 0x7fffffff` down to `hashCode()` — dropping the
+     * mask while the value actually wraps negative at runtime (→ bad array index). */
+    uint32_t h = 0;
     for (int64_t i = 0; i < s->len; i++) {
-        h = 31 * h + (int32_t)s->bytes[i];
+        h = 31u * h + (uint8_t)s->bytes[i];
     }
-    return h;
+    return (int32_t)h;
 }
 /* String.toString returns itself. */
 void *jrt_str_tostring(void *s) {
